Criando auxGetNomes que retona o nome do periodo e nome da montado
